backend/requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1

# AI/ML dependencies
google-generativeai==0.3.2
openai==1.3.8
transformers==4.36.2
torch==2.1.2
scikit-learn==1.3.2
numpy==1.25.2
pandas==2.1.4
accelerate==0.25.0
bitsandbytes==0.41.1

# NLP
spacy==3.7.2
nltk==3.8.1

# Medical libraries
pydicom==2.4.3
nibabel==5.2.0

# Video processing
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78

# Web scraping and requests
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
python-multipart==0.0.6
Pillow==10.1.0
pytesseract==0.3.10
python-magic==0.4.27
reportlab==3.6.12

# Performance monitoring
psutil==5.9.6

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0